refactor(apm): extract get_service_summary ranking into its own file
Keep apm_test.go under the 1k-line cap and drive Prom classes plus
sort keys from one catalog so unknown sort_by cannot silent-default.
